Simplify the LintError hierarchy
There is now no distinction between fixable- and non-fixable lint
errors. This is purely based on whether or not the linter provides a
fix.

This allows simplified generics, and the removal of a lot of classes.

This adds a restriction that now all auto-fixing linters must be able to
produce fixed file content from bad content, without modifying what's on
disk.