Allow overriding builtinLinters
This is useful if you want to run most of the project with all linters
by default, but for some subdirectory, run on a whitelist of linters.

For example, several Hack language tools should have generally unclean
test data, but those should still have the license header linter run on
them.