[PU] Starting linter for PU syntax (#216)
Creating linter to warn users if some of their files is using
reserved syntax for Pocket Universes.

Most of the syntax should just be erased from files right now.
It might happen that : and @ are close together, which will trigger
the linter, e.g. `function foo():@soft_annotation {...}`. In such cases,
the fix is to add a space in between