use XHP for hhast-inspect
This is mostly trivial, with a tiny bit of refactoring. Found a small
bug - pre/code open and close order were mismatched.

This should probably be split out to a separate repo; having HHAST
depend on xhp-lib isn't really acceptable, but neither is using HTML
strings when there's a better alternative.

fixes #323