Add simplified hhast-dump utility
This is primarily for debugging HHAST itself, for cross-checking with
`hh_parse --full-fidelity-json` - for example, to distinguish between
missing and 0-width nodes.

hh_parse remains the recommended way to canonically verify the parser's
behavior.

hhast-inspect (https://github.com/hhvm/hhast-inspect) remains the
recommended way to explore the AST of large files.

This distinction is useful when inspecting