Fix config with multiple overrides for a single file
This is a regression from
https://github.com/hhvm/hhast/commit/e6b406db4f4dbe0b7f6c53832d81e1e799faae87
that removes support for multiple overrides for a single
file.

This restores the support for multiple overrides for the run
config, but the `linterConfigs` section is still based on the
override found.

I manually tested this PR to make sure we don't regress
existing tests (and new tests are still passing).